% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{run_full_analysis}
\alias{run_full_analysis}
\title{Run full analysis: read CSVs, generate plots, optionally video}
\usage{
run_full_analysis(
  working_dir,
  output_dir = file.path(working_dir, "speed_out"),
  subject_name = "S01",
  enriched_dirs = NULL,
  generate_plots = TRUE,
  generate_video = FALSE,
  bids = FALSE,
  draw_gaze_path = TRUE,
  video_file = NULL
)
}
\arguments{
\item{working_dir}{Path to the data directory (can be un-enriched or BIDS root).}

\item{output_dir}{Path for saving results. Defaults to a 'speed_out' folder inside \code{working_dir}.}

\item{subject_name}{The subject identifier for annotating plots and files.}

\item{enriched_dirs}{A character vector of paths to "enriched" data folders, each corresponding to an AOI.}

\item{generate_plots}{Logical. If TRUE, summary plots will be generated.}

\item{generate_video}{Logical. If TRUE, a simple video overlay is created.}

\item{bids}{Logical. If TRUE, \code{working_dir} is interpreted as a BIDS root directory.}

\item{draw_gaze_path}{Logical. If TRUE, the video overlay will include a trail for the gaze.}

\item{video_file}{Optional path to a specific video file for the overlay.}
}
\value{
the output directory (invisible)
}
\description{
High-level pipeline
}